script tests verify if server is up, .gitignore and some specs

+++ tests.sh
@@ -2,11 +2,25 @@ git submodule init
 git submodule update 
 
 cd server/rest_in_practice/part_3/
-ruby script/server & 
-PID=$!
+
+server=$(curl http://localhost:3000/items.json -I -s | grep '200 OK')
+#server down
+PID=0
+if [[ $server == '' ]]; then
+  echo "turning server up"
+  ruby script/server & 
+  PID=$!
+  sleep 10
+else
+  echo "server already up"
+fi
+
 cd ../../../client
-sleep 10
 jspec run --rhino
-echo $PID
-kill -9 $PID
+
+if [[ $PID != 0 ]];then
+  echo $PID
+  kill -9 $PID
+fi
+
 cd ..
